Create UI-common
