+++ apps/browser/tsconfig.json
@@ -24,17 +24,18 @@
       "@bitwarden/generator-history": ["../../libs/tools/generator/extensions/history/src"],
       "@bitwarden/generator-legacy": ["../../libs/tools/generator/extensions/legacy/src"],
       "@bitwarden/generator-navigation": ["../../libs/tools/generator/extensions/navigation/src"],
+      "@bitwarden/importer/core": ["../../libs/importer/src"],
+      "@bitwarden/importer/ui": ["../../libs/importer/src/components"],
+      "@bitwarden/key-management": ["../../libs/key-management/src"],
+      "@bitwarden/key-management/angular": ["../../libs/key-management/src/angular"],
+      "@bitwarden/platform": ["../../libs/platform/src"],
+      "@bitwarden/send-ui": ["../../libs/tools/send/send-ui/src"],
+      "@bitwarden/tools-card": ["../../libs/tools/card/src"],
+      "@bitwarden/ui-core": ["./libs/ui/core/src"],
       "@bitwarden/vault-export-core": [
         "../../libs/tools/export/vault-export/vault-export-core/src"
       ],
       "@bitwarden/vault-export-ui": ["../../libs/tools/export/vault-export/vault-export-ui/src"],
-      "@bitwarden/importer/core": ["../../libs/importer/src"],
-      "@bitwarden/importer/ui": ["../../libs/importer/src/components"],
-      "@bitwarden/platform": ["../../libs/platform/src"],
-      "@bitwarden/send-ui": ["../../libs/tools/send/send-ui/src"],
-      "@bitwarden/tools-card": ["../../libs/tools/card/src"],
-      "@bitwarden/key-management": ["../../libs/key-management/src"],
-      "@bitwarden/key-management/angular": ["../../libs/key-management/src/angular"],
       "@bitwarden/vault": ["../../libs/vault/src"]
     },
     "plugins": [
