+++ apps/web/tsconfig.json
@@ -7,8 +7,8 @@
     "paths": {
       "@bitwarden/admin-console/common": ["../../libs/admin-console/src/common"],
       "@bitwarden/angular/*": ["../../libs/angular/src/*"],
-      "@bitwarden/auth/common": ["../../libs/auth/src/common"],
       "@bitwarden/auth/angular": ["../../libs/auth/src/angular"],
+      "@bitwarden/auth/common": ["../../libs/auth/src/common"],
       "@bitwarden/billing": ["../../libs/billing/src"],
       "@bitwarden/bit-common/*": ["../../bitwarden_license/bit-common/src/*"],
       "@bitwarden/common/*": ["../../libs/common/src/*"],
@@ -18,10 +18,6 @@
       "@bitwarden/generator-history": ["../../libs/tools/generator/extensions/history/src"],
       "@bitwarden/generator-legacy": ["../../libs/tools/generator/extensions/legacy/src"],
       "@bitwarden/generator-navigation": ["../../libs/tools/generator/extensions/navigation/src"],
-      "@bitwarden/vault-export-core": [
-        "../../libs/tools/export/vault-export/vault-export-core/src"
-      ],
-      "@bitwarden/vault-export-ui": ["../../libs/tools/export/vault-export/vault-export-ui/src"],
       "@bitwarden/importer/core": ["../../libs/importer/src"],
       "@bitwarden/importer/ui": ["../../libs/importer/src/components"],
       "@bitwarden/key-management": ["../../libs/key-management/src"],
@@ -29,6 +25,12 @@
       "@bitwarden/platform": ["../../libs/platform/src"],
       "@bitwarden/send-ui": ["../../libs/tools/send/send-ui/src"],
       "@bitwarden/tools-card": ["../../libs/tools/card/src"],
+      "@bitwarden/ui-common": ["../../libs/ui/common/src"],
+      "@bitwarden/ui-core": ["./libs/ui/core/src"],
+      "@bitwarden/vault-export-core": [
+        "../../libs/tools/export/vault-export/vault-export-core/src"
+      ],
+      "@bitwarden/vault-export-ui": ["../../libs/tools/export/vault-export/vault-export-ui/src"],
       "@bitwarden/vault": ["../../libs/vault/src"],
       "@bitwarden/web-vault/*": ["src/*"]
     }
