+++ libs/components/src/search/search.component.ts
@@ -9,10 +9,10 @@ import {
 } from "@angular/forms";
 
 import { isBrowserSafariApi } from "@bitwarden/platform";
+import { I18nPipe } from "@bitwarden/ui-common";
 
 import { InputModule } from "../input/input.module";
 import { FocusableElement } from "../shared/focusable-element";
-import { I18nPipe } from "../shared/i18n.pipe";
 
 let nextId = 0;
