+++ libs/components/src/shared/shared.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 
-import { I18nPipe } from "./i18n.pipe";
+import { I18nPipe } from "@bitwarden/ui-common";
 
 @NgModule({
   imports: [CommonModule, I18nPipe],
