+++ libs/components/tsconfig.json
@@ -20,7 +20,8 @@
     "lib": ["es2020", "dom"],
     "paths": {
       "@bitwarden/common/*": ["../common/src/*"],
-      "@bitwarden/platform": ["../platform/src"]
+      "@bitwarden/platform": ["../platform/src"],
+      "@bitwarden/ui-common": ["../ui/common/src"]
     },
     "plugins": [
       {
