+++ libs/importer/tsconfig.json
@@ -13,6 +13,7 @@
       "@bitwarden/generator-navigation": ["../tools/generator/extensions/navigation/src"],
       "@bitwarden/key-management": ["../key-management/src"],
       "@bitwarden/platform": ["../platform/src"],
+      "@bitwarden/ui-common": ["../ui/common/src"],
       "@bitwarden/vault-export-core": ["../tools/export/vault-export/vault-export-core/src"]
     }
   },
