+++ libs/ui/README
@@ -1,0 +1,5 @@
+# UI Foundation
+
+Core UI libraries maintained by the UI Foundation team.
+
+- _ui-common_: Low-level utilities for Angular applications.
