+++ libs/ui/common/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./di";
+export * from "./i18n.pipe";
