+++ tsconfig.eslint.json
@@ -26,16 +26,17 @@
       "@bitwarden/generator-history": ["./libs/tools/generator/extensions/history/src"],
       "@bitwarden/generator-legacy": ["./libs/tools/generator/extensions/legacy/src"],
       "@bitwarden/generator-navigation": ["./libs/tools/generator/extensions/navigation/src"],
-      "@bitwarden/vault-export-core": [".libs/tools/export/vault-export/vault-export-core/src"],
-      "@bitwarden/vault-export-ui": [".libs/tools/export/vault-export/vault-export-ui/src"],
       "@bitwarden/importer/core": ["./libs/importer/src"],
       "@bitwarden/importer/ui": ["./libs/importer/src/components"],
+      "@bitwarden/node/*": ["./libs/node/src/*"],
+      "@bitwarden/platform": ["./libs/platform/src"],
+      "@bitwarden/bit-common/*": ["./bitwarden_license/bit-common/src/*"],
       "@bitwarden/send-ui": [".libs/tools/send/send-ui/src"],
       "@bitwarden/tools-card": [".libs/tools/card/src"],
-      "@bitwarden/platform": ["./libs/platform/src"],
-      "@bitwarden/node/*": ["./libs/node/src/*"],
-      "@bitwarden/vault": ["./libs/vault/src"],
-      "@bitwarden/bit-common/*": ["./bitwarden_license/bit-common/src/*"]
+      "@bitwarden/ui-core": ["./libs/ui/core/src"],
+      "@bitwarden/vault-export-core": [".libs/tools/export/vault-export/vault-export-core/src"],
+      "@bitwarden/vault-export-ui": [".libs/tools/export/vault-export/vault-export-ui/src"],
+      "@bitwarden/vault": ["./libs/vault/src"]
     },
     "plugins": [
       {
