+++ tsconfig.json
@@ -18,9 +18,10 @@
     "paths": {
       "@bitwarden/admin-console/common": ["./libs/admin-console/src/common"],
       "@bitwarden/angular/*": ["./libs/angular/src/*"],
-      "@bitwarden/auth/common": ["./libs/auth/src/common"],
       "@bitwarden/auth/angular": ["./libs/auth/src/angular"],
+      "@bitwarden/auth/common": ["./libs/auth/src/common"],
       "@bitwarden/billing": ["./libs/billing/src"],
+      "@bitwarden/bit-common/*": ["./bitwarden_license/bit-common/src/*"],
       "@bitwarden/common/*": ["./libs/common/src/*"],
       "@bitwarden/components": ["./libs/components/src"],
       "@bitwarden/generator-components": ["./libs/tools/generator/components/src"],
@@ -28,19 +29,20 @@
       "@bitwarden/generator-history": ["./libs/tools/generator/extensions/history/src"],
       "@bitwarden/generator-legacy": ["./libs/tools/generator/extensions/legacy/src"],
       "@bitwarden/generator-navigation": ["./libs/tools/generator/extensions/navigation/src"],
-      "@bitwarden/vault-export-core": ["./libs/tools/export/vault-export/vault-export-core/src"],
-      "@bitwarden/vault-export-ui": ["./libs/tools/export/vault-export/vault-export-ui/src"],
       "@bitwarden/importer/core": ["./libs/importer/src"],
       "@bitwarden/importer/ui": ["./libs/importer/src/components"],
       "@bitwarden/key-management": ["./libs/key-management/src"],
       "@bitwarden/key-management/angular": ["./libs/key-management/src/angular"],
+      "@bitwarden/node/*": ["./libs/node/src/*"],
       "@bitwarden/platform": ["./libs/platform/src"],
       "@bitwarden/send-ui": ["./libs/tools/send/send-ui/src"],
       "@bitwarden/tools-card": ["./libs/tools/card/src"],
-      "@bitwarden/node/*": ["./libs/node/src/*"],
-      "@bitwarden/web-vault/*": ["./apps/web/src/*"],
+      "@bitwarden/ui-common": ["./libs/ui/common/src"],
+      "@bitwarden/ui-core": ["./libs/ui/core/src"],
+      "@bitwarden/vault-export-core": ["./libs/tools/export/vault-export/vault-export-core/src"],
+      "@bitwarden/vault-export-ui": ["./libs/tools/export/vault-export/vault-export-ui/src"],
       "@bitwarden/vault": ["./libs/vault/src"],
-      "@bitwarden/bit-common/*": ["./bitwarden_license/bit-common/src/*"]
+      "@bitwarden/web-vault/*": ["./apps/web/src/*"]
     },
     "plugins": [
       {
